fixed recent sonar warnings

+++ gemrb/core/Projectile.cpp
@@ -120,7 +120,7 @@ void Projectile::CreateAnimations(Animation **anims, const ResRef& bamres, int S
 	}
 
 	if((ExtFlags&PEF_CYCLE) && !Seq) {
-		Seq = RAND(size_t(0), Max - 1);
+		Seq = static_cast<int>(RAND(0UL, Max - 1));
 	}
 
 	//this hack is needed because bioware .pro files are sometimes
